+++ CHANGELOG.md
@@ -6,10 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Refresh button to sync and reload  wallet state
 
 ### Changed
 
 ### Fixed
+- Sign p2sh inputs with Ledger devices
 
 ## [1.0.5] - 2022-01-21
 ### Added